+++ lib/teamlab/modules/crm.rb
@@ -1,12 +1,14 @@
 require_relative 'crm/crm_cases'
 require_relative 'crm/crm_common'
 require_relative 'crm/crm_contacts'
+require_relative 'crm/crm_files'
 require_relative 'crm/crm_invoices'
 module Teamlab
   class Crm
     include CrmCases
     include CrmCommon
     include CrmContacts
+    include CrmFiles
     include CrmInvoices
 
     def initialize
@@ -321,30 +323,6 @@ module Teamlab
       @request.delete(['task', 'category', category_id.to_s])
     end
     # region Files
-
-    def get_root_folder_id
-      @request.get(%w[files root])
-    end
-
-    def get_file_list(entity_type, entity_id)
-      @request.get([entity_type.to_s, entity_id.to_s, 'files'])
-    end
-
-    def associate_file_with_entity(entity_type, entity_id, *fileids)
-      @request.post([entity_type, entity_id, 'files'], fileids: fileids.flatten)
-    end
-
-    def create_txt(entity_type, entity_id, title, content)
-      @request.post([entity_type.to_s, entity_id.to_s, 'files', 'text'], title: title, content: content)
-    end
-
-    def upload_file(entity_type, entity_id, file)
-      @request.post([entity_type.to_s, entity_id.to_s, 'files', 'upload'], somefile: File.new(file))
-    end
-
-    def delete_file(id)
-      @request.post(['files', id.to_s])
-    end
 
     def get_tags_for_entity_type(entity_type)
       @request.get([entity_type.to_s, 'tag'])