+++ Changelog.md
@@ -34,6 +34,7 @@
 * Fix `Crm#delete_case_group` request
 * Fix `Crm#set_rights_to_case` request
 * Fix `Crm#change_contact_photo` request
+* Fix `Crm#delete_file` request
 * Remove non-existing method `Crm#add_contact_address`
 * Remove non-existing method `Crm#update_contact_address`
 * Remove non-existing method `People#remind_password`